refactor: modifico beans de configuración a clase AppConfig
refactor: Se modifica la ubicación de los beans de configuración a la clase AppConfig
